requirements.txt
wtforms==2.3.3
parameterized==0.7.5
pytz==2020.5
workalendar==14.2.0
schedule==0.6.0
werkzeug==1.0.1
numpy==1.19.5
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
PyMySQL==1.0.1
SQLAlchemy==1.3.22
yahoo-fin==0.8.6
requests==2.25.1
requests-html==0.10.0
pandas==1.2.0
alpha-vantage==2.3.1
fake-useragent==0.1.11
lxml==4.6.2
SQLAlchemy-Utc==0.11.0
urllib3==1.25.9
trading-bot-predictor==0.0.1
tensorflow-gpu==2.4.1
keras==2.4.3
matplotlib==3.3.4
tensorflow==2.4.1
